+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
++ Frozen emote generation support for APNGs
+
 ## [1.0.0] - 2021-03-29
 
 ### Added